AdminBotService final test

+++ src/main/java/com/sarataza/atelieBot/Model/OrderEntity.java
@@ -12,7 +12,7 @@ public class OrderEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Integer number;
+    private Long number;
     private String works;
     private String LocalDate;
     private Boolean done;
